+++ R/HDI.R
@@ -32,28 +32,29 @@
 #' @references Kruschke JK. Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and Stan. 2nd edition. Academic Press, 2015
 #'
 #' @examples
-#' library(rstanarm)
-#' fit <- stan_glm(mpg ~ wt + am, data = mtcars, chains = 1)
-#' hdi(fit)
+#' if (require("rstanarm")) {
+#'   fit <- stan_glm(mpg ~ wt + am, data = mtcars, chains = 1)
+#'   hdi(fit)
 #'
-#' # fit logistic regression model
-#' fit <- stan_glm(
-#'   vs ~ wt + am,
-#'   data = mtcars,
-#'   family = binomial("logit"),
-#'   chains = 1
-#' )
-#' # compute hdi, transform on "odds ratio scale"
-#' hdi(fit, trans = exp)
+#'   # fit logistic regression model
+#'   fit <- stan_glm(
+#'     vs ~ wt + am,
+#'     data = mtcars,
+#'     family = binomial("logit"),
+#'     chains = 1
+#'   )
+#'   # compute hdi, transform on "odds ratio scale"
+#'   hdi(fit, trans = exp)
 #'
-#' # compute rope, on scale of linear predictor. finds proportion
-#' # of posterior distribution values between -1 and 1.
-#' rope(fit, rope = c(-1, 1))
+#'   # compute rope, on scale of linear predictor. finds proportion
+#'   # of posterior distribution values between -1 and 1.
+#'   rope(fit, rope = c(-1, 1))
 #'
-#' # compute rope, boundaries as "odds ratios". finds proportion of
-#' # posterior distribution values, which - after being exponentiated -
-#' # are between .8 and 1.25 (about -.22 and .22 on linear scale)
-#' rope(fit, rope = c(.8, 1.25), trans = exp)
+#'   # compute rope, boundaries as "odds ratios". finds proportion of
+#'   # posterior distribution values, which - after being exponentiated -
+#'   # are between .8 and 1.25 (about -.22 and .22 on linear scale)
+#'   rope(fit, rope = c(.8, 1.25), trans = exp)
+#' }
 #'
 #' @importFrom tibble as_tibble rownames_to_column
 #' @importFrom purrr map_dbl map_df